Refactoring Bugfix
