+++ ElevatedTrainStationTrack.csproj
@@ -77,14 +77,14 @@
     <Compile Include="NEXT\Texturing\TextureSet.cs" />
     <Compile Include="ObjectExtensions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Rail1LStation\Rail1L1SidedStation\Rail1L1SidedStationBuilder.cs" />
+    <Compile Include="Rail1LStation\Rail1L1SidedStation\Rail1LStationBuilder.cs" />
     <Compile Include="Rail1LStation\Rail1L2SidedStation\Rail1L2SidedStationBuilder.cs" />
-    <Compile Include="Rail1LStation\Rail1LStationBuilder.cs" />
-    <Compile Include="Rail1LStation\Rail1LStationBuilder.Texturing.cs" />
-    <Compile Include="Rail1L\Rail1L1W\Rail1L1WBuilder.cs" />
+    <Compile Include="Rail1LStation\Rail1LStationBuilderBase.cs" />
+    <Compile Include="Rail1LStation\Rail1LStationBuilderBase.Texturing.cs" />
+    <Compile Include="Rail1L\Rail1L1W\Rail1LBuilder.cs" />
     <Compile Include="Rail1L\Rail1L2W\Rail1L2WBuilder.cs" />
-    <Compile Include="Rail1L\Rail1LBuilder.cs" />
-    <Compile Include="Rail1L\Rail1LBuilder.Texturing.cs" />
+    <Compile Include="Rail1L\Rail1LBuilderBase.cs" />
+    <Compile Include="Rail1L\Rail1LBuilderBase.Texturing.cs" />
     <Compile Include="Rail2L1W\Rail2L1WBuilder.cs" />
     <Compile Include="SharedHelpers.cs" />
     <Compile Include="UI\ArrowsButtonSetup.cs" />
